Initial pass moving tests to PHPUnit with Junie.
The initial prompt rewrote the tests using normal phpunit style assertions instead of functional / CLI tests. With a second try being specific about that in the prompt it seemed to get the message.

Clear issues so far are that it used old unsupported versions of libraries.

The tests currently in `.github/workflows/test-disable-drupal-core-patches-level-check.yml` need to be rewritten to run with PHPUnit. The GitHub workflows should then run phpunit to test the library.

Modify the project to require phpunit as a dev dependency. Create new PHPUnit tests in a folder at `tests/src/Functional', with each test currently in the GitHub workflow created as a new test class. Create a basic `phpunit.xml.dist` file to support running these tests. Finally, update the github workflow to run the tests. Focus on porting the tests as directly as possible, without making unrelated improvements. The PHPUnit tests should use the Symfony process component to run the commands as currently shown in the workflow file. At the end, confirm that each line of shell scripting removed is accurately included in the new PHPUnit tests.